pyproject.toml
[project]
name = "hello-chatbot"
version = "0.1.0"
description = "Simple chatbot app"
authors = [
    {name = "shivaachukatla@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "simple-salesforce (>=1.12.6,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
